+++ packages/runtime-core/__tests__/helpers/useTemplateRef.spec.ts
@@ -3,6 +3,7 @@ import {
   h,
   nextTick,
   nodeOps,
+  onMounted,
   ref,
   render,
   useTemplateRef,
@@ -249,6 +250,52 @@ describe('useTemplateRef', () => {
 
       expect('target is readonly').not.toHaveBeenWarned()
       expect(foo!.value).toBe(root.children[0])
+    } finally {
+      __DEV__ = true
+    }
+  })
+
+  // #12749
+  test(`don't update setup ref for useTemplateRef key`, () => {
+    let foo: ShallowRef
+    const Comp = {
+      setup() {
+        foo = useTemplateRef('bar')
+        const bar = ref(null)
+        onMounted(() => {
+          expect(bar.value).toBe(null)
+        })
+        return { bar }
+      },
+      render() {
+        return h('div', { ref: 'bar' })
+      },
+    }
+    const root = nodeOps.createElement('div')
+    render(h(Comp), root)
+    expect(foo!.value).toBe(root.children[0])
+  })
+
+  test(`don't update setup ref for useTemplateRef key (compiled in prod mode)`, () => {
+    __DEV__ = false
+    try {
+      let foo: ReturnType<typeof ref>
+      let fooRef: ShallowRef
+      const Comp = {
+        setup() {
+          foo = ref('hello')
+          fooRef = useTemplateRef('foo')
+          return { foo }
+        },
+        render() {
+          return h('input', { ref: foo, ref_key: 'foo' })
+        },
+      }
+      const root = nodeOps.createElement('div')
+      render(h(Comp), root)
+
+      expect(foo!.value).toBe('hello')
+      expect(fooRef!.value).toBe(root.children[0])
     } finally {
       __DEV__ = true
     }