+++ packages/runtime-core/src/components/Teleport.ts
@@ -169,9 +169,13 @@ export const TeleportImpl = {
         updateCssVars(n2, true)
       }
 
-      if (isTeleportDeferred(n2.props)) {
+      if (
+        isTeleportDeferred(n2.props) ||
+        (parentSuspense && parentSuspense.pendingBranch)
+      ) {
         n2.el!.__isMounted = false
         queuePostRenderEffect(() => {
+          if (n2.el!.__isMounted !== false) return
           mountToTarget()
           delete n2.el!.__isMounted
         }, parentSuspense)
@@ -179,7 +183,17 @@ export const TeleportImpl = {
         mountToTarget()
       }
     } else {
-      if (isTeleportDeferred(n2.props) && n1.el!.__isMounted === false) {
+      // update content
+      n2.el = n1.el
+      n2.targetStart = n1.targetStart
+      const mainAnchor = (n2.anchor = n1.anchor)!
+      const target = (n2.target = n1.target)!
+      const targetAnchor = (n2.targetAnchor = n1.targetAnchor)!
+
+      // Target mounting may still be pending because of deferred teleport or a
+      // parent suspense buffering post-render effects. In that case, defer the
+      // teleport patch itself until the pending mount effect has run.
+      if (n1.el!.__isMounted === false) {
         queuePostRenderEffect(() => {
           TeleportImpl.process(
             n1,
@@ -196,12 +210,6 @@ export const TeleportImpl = {
         }, parentSuspense)
         return
       }
-      // update content
-      n2.el = n1.el
-      n2.targetStart = n1.targetStart
-      const mainAnchor = (n2.anchor = n1.anchor)!
-      const target = (n2.target = n1.target)!
-      const targetAnchor = (n2.targetAnchor = n1.targetAnchor)!
       const wasDisabled = isTeleportDisabled(n1.props)
       const currentContainer = wasDisabled ? container : target
       const currentAnchor = wasDisabled ? mainAnchor : targetAnchor
@@ -412,12 +420,30 @@ function hydrateTeleport(
     optimized: boolean,
   ) => Node | null,
 ): Node | null {
-  function hydrateDisabledTeleport(
-    node: Node,
-    vnode: VNode,
-    targetStart: Node | null,
-    targetAnchor: Node | null,
+  // lookahead until we find the target anchor
+  // we cannot rely on return value of hydrateChildren() because there
+  // could be nested teleports
+  function hydrateAnchor(
+    target: TeleportTargetElement,
+    targetNode: Node | null,
   ) {
+    let targetAnchor = targetNode
+    while (targetAnchor) {
+      if (targetAnchor && targetAnchor.nodeType === 8) {
+        if ((targetAnchor as Comment).data === 'teleport start anchor') {
+          vnode.targetStart = targetAnchor
+        } else if ((targetAnchor as Comment).data === 'teleport anchor') {
+          vnode.targetAnchor = targetAnchor
+          target._lpa =
+            vnode.targetAnchor && nextSibling(vnode.targetAnchor as Node)
+          break
+        }
+      }
+      targetAnchor = nextSibling(targetAnchor)
+    }
+  }
+
+  function hydrateDisabledTeleport(node: Node, vnode: VNode) {
     vnode.anchor = hydrateChildren(
       nextSibling(node),
       vnode,
@@ -427,8 +453,6 @@ function hydrateTeleport(
       slotScopeIds,
       optimized,
     )
-    vnode.targetStart = targetStart
-    vnode.targetAnchor = targetAnchor
   }
 
   const target = (vnode.target = resolveTarget<Element>(
@@ -443,33 +467,22 @@ function hydrateTeleport(
       (target as TeleportTargetElement)._lpa || target.firstChild
     if (vnode.shapeFlag & ShapeFlags.ARRAY_CHILDREN) {
       if (disabled) {
-        hydrateDisabledTeleport(
-          node,
-          vnode,
-          targetNode,
-          targetNode && nextSibling(targetNode),
-        )
+        hydrateDisabledTeleport(node, vnode)
+        hydrateAnchor(target as TeleportTargetElement, targetNode)
+        if (!vnode.targetAnchor) {
+          prepareAnchor(
+            target,
+            vnode,
+            createText,
+            insert,
+            // if target is the same as the main view, insert anchors before current node
+            // to avoid hydrating mismatch
+            parentNode(node)! === target ? node : null,
+          )
+        }
       } else {
         vnode.anchor = nextSibling(node)
-
-        // lookahead until we find the target anchor
-        // we cannot rely on return value of hydrateChildren() because there
-        // could be nested teleports
-        let targetAnchor = targetNode
-        while (targetAnchor) {
-          if (targetAnchor && targetAnchor.nodeType === 8) {
-            if ((targetAnchor as Comment).data === 'teleport start anchor') {
-              vnode.targetStart = targetAnchor
-            } else if ((targetAnchor as Comment).data === 'teleport anchor') {
-              vnode.targetAnchor = targetAnchor
-              ;(target as TeleportTargetElement)._lpa =
-                vnode.targetAnchor && nextSibling(vnode.targetAnchor as Node)
-              break
-            }
-          }
-          targetAnchor = nextSibling(targetAnchor)
-        }
-
+        hydrateAnchor(target as TeleportTargetElement, targetNode)
         // #11400 if the HTML corresponding to Teleport is not embedded in the
         // correct position on the final page during SSR. the targetAnchor will
         // always be null, we need to manually add targetAnchor to ensure
@@ -492,7 +505,9 @@ function hydrateTeleport(
     updateCssVars(vnode, disabled)
   } else if (disabled) {
     if (vnode.shapeFlag & ShapeFlags.ARRAY_CHILDREN) {
-      hydrateDisabledTeleport(node, vnode, node, nextSibling(node))
+      hydrateDisabledTeleport(node, vnode)
+      vnode.targetStart = node
+      vnode.targetAnchor = nextSibling(node)
     }
   }
   return vnode.anchor && nextSibling(vnode.anchor as Node)
@@ -535,6 +550,7 @@ function prepareAnchor(
   vnode: TeleportVNode,
   createText: RendererOptions['createText'],
   insert: RendererOptions['insert'],
+  anchor: RendererNode | null = null,
 ) {
   const targetStart = (vnode.targetStart = createText(''))
   const targetAnchor = (vnode.targetAnchor = createText(''))
@@ -544,8 +560,8 @@ function prepareAnchor(
   targetStart[TeleportEndKey] = targetAnchor
 
   if (target) {
-    insert(targetStart, target)
-    insert(targetAnchor, target)
+    insert(targetStart, target, anchor)
+    insert(targetAnchor, target, anchor)
   }
 
   return targetAnchor